pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-deepseek>=1.0.0",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
